third_party/WebKit/LayoutTests/css3/device-adapt/viewport-insert-rule-after-expected.txt

ALERT: viewport size 1000x1100 scale 0.32 with limits [0.32, -1] and userScalable true
